+++ renku/core/util/git.py
@@ -627,7 +627,7 @@ def clone_renku_repository(
     install_githooks=False,
     install_lfs=True,
     skip_smudge=True,
-    recursive=True,
+    recursive=False,
     progress=None,
     config: Optional[dict] = None,
     raise_git_except=False,
@@ -646,7 +646,7 @@ def clone_renku_repository(
         install_githooks: Whether to install git hooks (Default value = False).
         install_lfs: Whether to install Git LFS (Default value = True).
         skip_smudge: Whether to pull files from Git LFS (Default value = True).
-        recursive: Whether to clone recursively (Default value = True).
+        recursive: Whether to clone recursively (Default value = False).
         progress: The GitProgress object (Default value = None).
         config(Optional[dict], optional): Set configuration for the project (Default value = None).
         raise_git_except: Whether to raise git exceptions (Default value = False).
@@ -712,7 +712,7 @@ def clone_repository(
     install_githooks=True,
     install_lfs=True,
     skip_smudge=True,
-    recursive=True,
+    recursive=False,
     depth=None,
     progress: Optional[git.RemoteProgress] = None,
     config: Optional[dict] = None,
@@ -730,7 +730,7 @@ def clone_repository(
         install_githooks: Whether to install git hooks (Default value = True).
         install_lfs: Whether to install Git LFS (Default value = True).
         skip_smudge: Whether to pull files from Git LFS (Default value = True).
-        recursive: Whether to clone recursively (Default value = True).
+        recursive: Whether to clone recursively (Default value = False).
         depth: The clone depth, number of commits from HEAD (Default value = None).
         progress: The GitProgress object (Default value = None).
         config(Optional[dict], optional): Set configuration for the project (Default value = None).
@@ -760,7 +760,7 @@ def clone_repository(
 
         return errors.GitError(message)
 
-    def clean_directory():
+    def clean_directory(clean: bool):
         if not clean or not path:
             return
         try:
@@ -791,10 +791,10 @@ def clone_repository(
                     pass
             else:
                 # NOTE: not same remote, so don't reuse
-                clean_directory()
+                clean_directory(clean=clean)
                 return None
         except errors.GitError:  # NOTE: Not a git repository, remote not found, or checkout failed
-            clean_directory()
+            clean_directory(clean=clean)
         else:
             return repository
 
@@ -828,6 +828,9 @@ def clone_repository(
             if raise_git_except:
                 raise
             raise error_from_progress(progress, url)
+
+        # NOTE: Delete the partially-cloned repository
+        clean_directory(clean=True)
 
         # NOTE: clone without branch set, in case checkout_revision was not a branch or a tag but a commit
         try: