+++ renku/ui/service/controllers/datasets_edit.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2020 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,7 +17,7 @@
 from typing import Dict, List, Union, cast
 
 from renku.command.dataset import edit_dataset_command
-from renku.core.dataset.request_model import ImageRequestModel
+from renku.core.image import ImageObjectRequest
 from renku.core.util.metadata import construct_creators
 from renku.domain_model.constant import NO_VALUE, NoValueType
 from renku.domain_model.provenance.agent import Person
@@ -39,8 +38,8 @@ class DatasetsEditCtrl(ServiceCtrl, RenkuOpSyncMixin):
 
     def __init__(self, cache, user_data, request_data, migrate_project=False):
         """Construct a datasets edit list controller."""
-        self.ctx = cast(Dict, DatasetsEditCtrl.REQUEST_SERIALIZER.load(request_data))
-        self.ctx["commit_message"] = f"{MESSAGE_PREFIX} dataset edit {self.ctx['name']}"
+        self.ctx = cast(Dict, self.REQUEST_SERIALIZER.load(request_data))
+        self.ctx["commit_message"] = f"{MESSAGE_PREFIX} dataset edit {self.ctx['slug']}"
 
         super().__init__(cache, user_data, request_data, migrate_project=migrate_project)
 
@@ -61,7 +60,7 @@ class DatasetsEditCtrl(ServiceCtrl, RenkuOpSyncMixin):
             set_url_for_uploaded_images(images=images, cache=self.cache, user=self.user)
 
             images = [
-                ImageRequestModel(
+                ImageObjectRequest(
                     content_url=img["content_url"],
                     position=img["position"],
                     mirror_locally=img.get("mirror_locally", False),
@@ -78,10 +77,10 @@ class DatasetsEditCtrl(ServiceCtrl, RenkuOpSyncMixin):
         else:
             creators = NO_VALUE
 
-        if "title" in self.ctx:
-            title = self.ctx.get("title")
+        if "name" in self.ctx:
+            name = self.ctx.get("name")
         else:
-            title = NO_VALUE
+            name = NO_VALUE
 
         if "description" in self.ctx:
             description = self.ctx.get("description")
@@ -111,8 +110,8 @@ class DatasetsEditCtrl(ServiceCtrl, RenkuOpSyncMixin):
             .with_commit_message(self.ctx["commit_message"])
             .build()
             .execute(
-                self.ctx["name"],
-                title=title,
+                self.ctx["slug"],
+                name=name,
                 description=description,
                 creators=creators,
                 keywords=keywords,
@@ -143,4 +142,4 @@ class DatasetsEditCtrl(ServiceCtrl, RenkuOpSyncMixin):
             "remote_branch": remote_branch,
         }
 
-        return result_response(DatasetsEditCtrl.RESPONSE_SERIALIZER, response)
+        return result_response(self.RESPONSE_SERIALIZER, response)