+++ tests/cli/test_init.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2018-2023- Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -471,3 +470,45 @@ def test_init_with_description(isolated_runner, template):
     result = isolated_runner.invoke(cli, ["graph", "export", "--strict"])
     assert 0 == result.exit_code, format_result_exception(result)
     assert "my project description" in result.output
+
+
+@pytest.mark.parametrize(
+    "image",
+    [Path(__file__).parent / ".." / "data" / "renku.png", "https://en.wikipedia.org/static/images/icons/wikipedia.png"],
+)
+def test_init_with_image(isolated_runner, template, image):
+    """Test project initialization with image."""
+    result = isolated_runner.invoke(cli, ["init", "--image", image, "new-project", "-t", template["id"]])
+
+    assert 0 == result.exit_code, format_result_exception(result)
+
+    project = Database.from_path(Path("new-project") / ".renku" / "metadata").get("project")
+
+    assert (Path("new-project") / ".renku" / "images" / "project" / "0.png").exists()
+    assert ".renku/images/project/0.png" == project.image.content_url
+    assert 0 == project.image.position
+
+
+@pytest.mark.parametrize(
+    "image, error",
+    [
+        ("non-existing.png", "Image with local path"),
+        ("/non-existing", "Image with local path"),
+        ("https://example.com/non-existing.png", "Cannot download image with url"),
+    ],
+)
+def test_init_with_non_existing_image(isolated_runner, template, image, error):
+    """Test project initialization fails when the image doesn't exist."""
+    result = isolated_runner.invoke(cli, ["init", "--image", image, "new-project", "-t", template["id"]])
+
+    assert 1 == result.exit_code, format_result_exception(result)
+    assert error in result.output
+
+
+@pytest.mark.parametrize("image", [Path(__file__).parent / ".." / "data" / "workflow-file.yml", "/etc/passwd"])
+def test_init_with_non_image_files(isolated_runner, template, image):
+    """Test project initialization fails when the file isn't an image format."""
+    result = isolated_runner.invoke(cli, ["init", "--image", image, "new-project", "-t", template["id"]])
+
+    assert 1 == result.exit_code, format_result_exception(result)
+    assert "isn't a valid image file" in result.output