+++ tests/cli/test_template.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2018-2023- Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -91,7 +90,7 @@ def test_template_show(isolated_runner):
         result = isolated_runner.invoke(cli, command + ["R-minimal"])
 
         assert 0 == result.exit_code, format_result_exception(result)
-        assert re.search("^Name: Basic R (.*) Project$", result.output, re.MULTILINE) is not None
+        assert re.search("^Name: R (.*) Project$", result.output, re.MULTILINE) is not None
     finally:
         sys.argv = argv
 
@@ -101,7 +100,7 @@ def test_template_show_no_id(runner, project):
     result = runner.invoke(cli, ["template", "show"])
 
     assert 0 == result.exit_code, format_result_exception(result)
-    assert re.search("^Name: Basic Python (.*) Project$", result.output, re.MULTILINE) is not None
+    assert re.search("^Name: Python (.*) Project$", result.output, re.MULTILINE) is not None
 
 
 def test_template_show_no_id_outside_project(isolated_runner):